Backend Completed with Authorization

+++ MERN_Stack/backend/controller/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')// package error handler replacement of try catch
 const { Mongoose } = require('mongoose')
 const Goal = require('../model/goalModel')
+const User = require('../model/userModel')
 
 // description = get goals
 // route = Get /
 // access private
 const getGoal = asyncHandler(async (req, res) => {
-    const goal = await Goal.find()
+    const goal = await Goal.find({user:req.user.id})
     res.status(200).json(goal)
 })
 
@@ -18,7 +19,8 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Please add a text field in postman')
     }
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user.id
     })
     res.status(200).json(goal)
 })
@@ -29,6 +31,21 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error("Goal not found")
     }
+
+    const user = await User.findById(req.user.id)
+
+    // checking for user existence
+    if (!user) {
+        res.status(401)
+        throw new Error("User Not Found")
+    }
+
+    // matching with logged in user so to update only those who has access to
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error("User Not Authorized ")
+    };
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
     })
@@ -41,6 +58,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error("Goal not found")
     }
+
+    const user = await User.findById(req.user.id)
+
+    // checking for user existence
+    if (!user) {
+        res.status(401)
+        throw new Error("User Not Found")
+    }
+
+    // matching with logged in user so to delete only those who has access to
+    if (goal.user.toString() != user.id) {
+        res.status(401)
+        throw new Error("User Not Authorized ")
+    }
+
     await Goal.findByIdAndRemove(req.params.id)
 
     res.status(200).json({id: req.params.id})
